requirements.txt
fastapi[standard]
pydantic==2.11.3
pydantic_core==2.33.1
pydantic_settings==2.8.1
sqlmodel==0.0.24
psycopg2-binary==2.9.10
argon2-cffi==19.1.0
python-jose==3.3.0